requirements.txt
discord.py
python-dotenv
google-generativeai

